# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "This is a Graduation Application Management System.  It is implemented on NodeJs with REST API Architecture. You can find\nout source code on\n[Github](https://github.com/microbraino/gradapp)\n",
        "version": "1.0.0",
        "title": "Graduate Application",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "abdulkadir.kahraman@outlook.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "description": "Heroku Server",
            "url": "http://commerchant.herokuapp.com"
        }
    ],
...
    "paths": {
...
        "/applications": {
            "get": {
                "tags": [
                    "applications"
                ],
                "x-roles": [
                    "applicant"
                ],
                "summary": "get application of current logged in account session",
                "operationId": "getApplication",
                "requestBody": {
                    "$ref": "#/components/requestBodies/updateApplication"
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "applications"
                ],
                "x-roles": [
                    "applicant"
                ],
                "summary": "Update some permited field of application of current logged in account session",
                "operationId": "updateApplication",
                "requestBody": {
                    "$ref": "#/components/requestBodies/updateApplication"
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "patch": {
                "tags": [
                    "applications"
                ],
                "x-roles": [
                    "applicant"
                ],
                "summary": "Update application status of current logged in account session",
                "operationId": "setStatus",
                "requestBody": {
                    "$ref": "#/components/requestBodies/setStatus"
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
...
    "components": {
...
        "requestBodies": {
...
            "setStatus": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "enum": [
                                        "created",
                                        "edited",
                                        "submited",
                                        "checked",
                                        "updateRequested",
                                        "updated",
                                        "confirmed",
                                        "rejected",
                                        "assessed",
                                        "accepted",
                                        "announced"
                                    ]
                                }
                            }
                        }
                    }
                },
                "required": true
            },
            "updateApplication": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "photo": {
                                    "type": "string"
                                },
                                "undergradTranscript": {
                                    "type": "string",
                                    "example": "example_document_id"
                                },
                                "alesResult": {
                                    "type": "string",
                                    "example": "example_document_id"
                                },
                                "englishExamResult": {
                                    "type": "string",
                                    "example": "example_document_id"
                                },
                                "referenceLetters": {
                                    "type": "string",
                                    "example": "example_document_id"
                                },
                                "statementOfPurpose": {
                                    "type": "string",
                                    "example": "example_document_id"
                                },
                                "nationality": {
                                    "type": "string",
                                    "example": "Turk"
                                },
                                "passportCopy": {
                                    "type": "string",
                                    "example": "example_document_id"
                                },
                                "doesWork": {
                                    "type": "boolean",
                                    "example": "false"
                                },
                                "permissionLetter": {
                                    "type": "string",
                                    "example": "example_document_id"
                                },
                                "masterTranscript": {
                                    "type": "string",
                                    "example": "example_document_id"
                                }
                            }
                        }
                    }
                },
                "required": true
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}